Super early preliminary
This is a preliminary website, not nearly completed. I wanted to try
using embedded fonts but couldn't get them to work, so it would be nice
if someone could get the h1 and h2 to display with FifteenNarrow.ttf. I
wanted to create a sort of terminal thing, so all the text is in green
on a black background. Monospace sort of looks terminalish, but
FifteenNarrow is the font I'm looking to accomplish this.